+++ govtool/frontend/src/utils/adaFormat.ts
@@ -8,6 +8,20 @@ export const correctAdaFormat = (lovelace: number | undefined) => {
   return 0;
 };
 
+export const correctVoteAdaFormat = (
+  lovelace: number | undefined,
+  locale: string | undefined = undefined,
+) => {
+  if (lovelace) {
+    const ada = lovelace / LOVELACE;
+    return ada.toLocaleString(locale, {
+      minimumFractionDigits: 3,
+      maximumFractionDigits: 3,
+    });
+  }
+  return "0,000";
+};
+
 export const correctDRepDirectoryFormat = (lovelace: number | undefined) => {
   if (lovelace) {
     return Number((lovelace / LOVELACE).toFixed(0))?.toLocaleString("en-US");